Fixed error with function definition arguments.
Function definition arguments were overlapping with the outer context due to incorrect ordering of appended environments. Fixed this issue for Function Definitions and card config ContextForEach objects, and checked in several other likely places. This issue may occur in other places in the codebase, and is one to watch out for.