Working version of CardConfig documentation.
There is now a working and complete version of the CardConfig/CardSetConfig documentation for the editor. Configs can now have names and descriptions, and these will be displayed in the documentation window entries. Definitions still need some corrections to the way they are displayed in the editor. No changes yet to tooltip content for card configuration files.