Implement (partial) horizontal font kerning.
Font kerning has been implemented in GeboPDF and the necessary changes made in SharpSheets and SharpEditor. Currently only the 'kern' table is being used, and only for horizontal kerning (although all 'kern' data is being read). Only the Windows version is currently accepted for the 'kern' table (version 0, and subtable formats 0 and 2). The GPOS table should also be implemented in future. Vertical kerning will require rewrites to the other text handling algorithms for dealing with vertical writing systems, if we get there.