Added OpenType functionality to fonts in GeboPDF and SharpSheets.
The OpenType functionality has now been incorporated into the PdfType0Font class, and implemented in GraphicsStream. Still need to include placement data in glyph arrangement. These updates have been incorporated into SharpSheets, and improvements have been made in the handling of fonts in the SharpSheets library (all fonts are now backed by PdfType0Font objects). As a result, the Liberation Sans font family has been added to the SharpSheets library to use as a default font in both the designer and generated documents. The font usage data (e.g. ligature encodings) still needs to be properly implemented, and a mechanism for specifying OpenType features in SharpSheets needs to be created.