Implement FontGlyphUsage data into ToUnicode creation.
The FontGlyphUsage class has been incorporated into the CMap creation for the ToUnicode cmap. The process of tracking glyph usage is perhaps a little brittle currently, and could possibly do with improving in the future. OpenTypeLayoutTags has been moved to an argument of CIDFontFactory.CreateFont methods, and must now be dealt with on the SharpSheets side.