Refactor font arguments.
Font arguments are now condensed into a single object, which be unpacked into all the necessary properties when used configuration files. This also allows for better unpacking of the tags and paths, such that tags can be specified for individual fonts within a typeface, when the typeface is specified as a single font family name. These changes necessitated an update to the ISharpArgSupplemented functionality, such that it now allows for optional/nullable first arguments.