Added error popup to SharpEditor.
There is now an up-arrow on the error status bar, which opens a popup containing a scrollable list of the document errors. Double-clicking on one of these entries scrolls the cursor to the location of the error (if the error has a location). Support for errors without locations has also been added (but needs testing for completeness).