Better error handing for SharpPageList.
A try-catch has been inserted in the SharpPageList.DrawTo method, to capture any drawing errors which happen at the page level, before the SharpWidget error handling can take over. There was an edge case where page-level errors (for example, when applying the page WidgetSetup) were not being captured properly, and so could not be displayed clearly in the editor.