First attempt at Font Subsetting.
Initial effort to subset font tables (not glyphs). Only the tables required for PDF will be written into the file (as specified in the PDF specification). Currently being generous with 'cmap' and tables for OpenType (i.e. 'CFF ') fonts, just to be safe. Created a FontFileWriter class to hold logic for writing tables to an output stream (this should be extended for more functionality later).