Add visualisation of OpenType features in documentation.
OpenType substitutions are now visually represented in the font documentation pages. A special "OpenType Features" page has ben created to show the list of substititions that can be made with each feature. Different arrows are used to indicate whether the substitutions are automatic, require a choice (although this behaviour is not yet implemented), or are examples from a larger possible set of substitutions (with an "e.g."). Functionality for getting example substitutions has been added to GeboPDF to support this (this is a little messy at present, and could be cleaner).