Improvements to OpenType positioning.
Mark-to-Base positioning now properly looks backwards in the sequence, and Mark-to-Mark positioning ignores any existing placement on the attaching mark. GraphicsStream has been updated to reduce unnecessary instructions in the PDF file, and had a bug with final text rise timing fixed. String parsing in SharpSheets has been updated to include Unicode normalization, which is needed for correctly grouping character combining classes (currently this is using FormKD, as this proved to work best, but the best method needs determining properly).