Add full support for fonts in template directory.
Fonts in the template directory will now be collected and presented to the user in the documentation along with system fonts. Support has also been added for importing font files by dropping (which possibly should be extended to more file types). This is supported by the option to add custom font source directories to FontPathRegistry. Also some small fixes to template importing and registry handling.