Improve autocomplete and tooltip for namedChild and fallback types.
Autocompletion will now include "&" for named child elements. Additionally, hovering over empty space above the first element will display a tooltip for the fallback element type in Sharp config files (this may prove a little annoying, and may need to appear in more specific circumstances). Improvements have also been made to the autocompletion window, as it will now collapse options with the same text content and type into single entries. This necessitated a refactoring of the multiple arguments tooltip, and some improvements to the IEquatable interfaces for certain types (and fixing a bug in the DocumentationString implementation).