Add example value syntax to card constant definition.
Constant definitions can now accept an example value, which can be used as a substitute when doing a dry-run (which has been implemented), and when populating default values when adding a section to a card subject file (not yet implemented). The DryRun changes are a little hacky, and could possibly be streamlined.