Fix bug in PDF drawing of OpenType features.
OpenType features were not being rendered properly due to an incorrect bypass in the GraphicsStream text layout method, and the wrong font being selected when using two fonts with matching origins but different OpenType feature sets. This has necessitated redefining how fonts are selected and combined in the ResourceDictionary and PdfDocumentWriter classes, along with the necessary changes to the PdfFont classes. The string normalisation when parsing has also been adjusted to give more appropriate results.