Fix bug with regex definition type and escaped characters.
The regex definition type now parses the incoming text using StringParsing before comparing with the provided regex. This removes the need to put "\\" in the regex if asterisk or underscore characters are required for rich strings. Additionally, the card subject parser has been updated to not parse line terminators for initial setup entries, and to allow the use of three "\" to indicate that the line should end with a "\" character (this may not be the best behaviour, needs considering).