Changes to context usage tracking and implement config foreach.
foreach properties can now be used in sheet config files. This has necessitated the rewriting of some IContext functionality (same overall result, but better separation of concerns). Usage tracking has been moved to its own class, VisitTrackingContext. InterpolatedContext has been moved to the Evaluations namespace. The code for tracking line ownership has also been streamlined, with LineOwnership and IReadOnlyLineOwnership types created to handle this logic.

The foreach properties are integrated into the parsing, and repeated elements highlight appropriately in the editor. However, there is no documentation for the foreach property (either written in the docs, or displayed to the user). The card type collection functionality for adding these arguments could be adapted for this purpose.