Implement font outline reading (glyf and CFF).
Implementation for reading glyph outlines from glyf and CFF tables. These can then be used to display glyphs within the designer canvas, without the need for the WPF GlyphTypeface class. In future, this could also be used as a basis to implement variable font rendering, and font subsetting. Necessary updates have been made to SharpWPFDrawingCanvas, TypefaceGrouping, and the FontGraphicsRegistry.