Create SharpGeometryDrawingCanvas, other minor fixes.
Created an ISharpCanvas class that can perform most calculations off the UI thread, and then can separately have its output converted to Avalonia UI elements on the UI thread at a later point. This is to solve a noticeable lag that occured during layout (which was occuring on the UI thread due to Avalonia constraints). There is still a slight lag when first drawing, however. This could maybe be mitigated by creating a custom deferred drawing canvas that stores Skia canvas operations.