Add Markup EntriedShape class.
This necessitated changing the way EntryCount is made accessible. Now it is a method that takes in the graphics state and rect, so that these values can be used in calculations for Markup shapes. DrawableDivElement has been cleaned up a little to better separate incoming element types, and disambiguate AreaElements (so that their names can now be calculated, so allow for variable number of entries in entried shapes).