Add DPI scaling to CanvasViewer.
CanvasViewer now has a DPI attribute which can be adjusted to change the scaling of the canvas content. A default value of 96 has been used. Unsure how this will interact with screens of different resolutions, as Avalonia does already attempt to account for this (apparently). However, the adjustment will still allow fine-tuning, even if it does not match up numerically on all systems (needs further testing to be certain of behaviour).