Custom ItemDragBehavior implementation for WrapPanel ItemsPresenter.
The Avalonia.Xaml.Interactivity default implementation of ItemDragBehavior replaces (in the associated Style) the WrapPanel for the TabControl item headers with a StackPanel. The changes to ItemDragBehavior.PointerMoved allow for multiple rows of headers, and the associated Style has been altered so as not to replace the ItemsPresenter panel. The implementation is a little messy, and could possibly be streamlined, but functions adequately so far. Other than PointerMoved, the class has not been changed substantially.