Improve settings window behaviours.
The TopMost="True" for SettingsWindow has been removed, in favour of opening it as a dialogue of the main window. This fixes the issue of message boxes appearing behind the settings window, but keeps the settings window on top of the main one (and allows for switching between applications while settings is open). Not perfect,  but better. Additionally, the starting location for the template directory picker has been corrected to the parent directory of the template directory itself.