Remove NotImplementedException to prevent annoying Avalonia error.
Avalonia requires a parameterless constructor for Controls, otherwise you get a warning. The SharpEditorWindow had a NotImplementedException in its  parameterless constructor, which was causing a separate error whenever the AXAML file was opened in the IDE. The associated attribute has been made nullable in order to allow for no value at initialisation, and the associated references adjusted as appropriate. Hopefully now the error will go away.