Custom styling for dialogues.
A custom MsBoxSharpView has been created, with associated types, which uses vector icons specified in IconStyles. Additionally, MessageBoxes has been moved into a Dialogues namespace (Windows is too cluttered at the moment). This has necessitated some adjustment to the MessageBox interface (including making the dialogues an actual dialogue popup), and to the Generator class (which now fires an event when an error is encountered, rather than directly show a MessageBox. Currently only a limited number of icons have been implemented for the dialogue, but this can be easily extended in MsBoxSharpView, MsBoxSharpViewModel, and IconStyles.