Fix editor ToolTip open/close behaviour.
Using IsVisible along with SetOpen was causing rendering issues on Linux, as the ToolTip was attempting to open when it had no content, and was flashing a black box on the screen. By instead caching the ToolTip object and using SetTip, to set and unset the TextEditorToolTip.Tip, this issue is resolved. It's a bit hacky, but it works (it's easier to setup the ToolTip object from AXAML, rather than build it in codebehind, but the caching is a little off).