Fix unused highlighting and move color to SharpEditorPalette.
The "unused" syntax highlighting was highlighting too much of the document, particularly fully commented lines. This has been fixed, along with moving the "unused" color into the highlighting colors, so that the font style and weight can also be adjusted. The parsing state rendering code has been adjusted to make use of this, and up update properly when the highlighting colors are changed.