Allow Markup URLs to evaluate from anywhere in pattern.
Previously, Markup URLs could only find elements at the same level in the div element tree, rather than any element in the pattern. While this is safer in many circumstances, it is quite limiting in many ways. Now an element may be found anywhere inside the same pattern - and will use the width, etc., of where it is called from, not where it is defined. Further checks and error handling may be needed to make this smoother in practice, but the core works for now.