Fix bug in glyph reading with off curve start/end points.
When a later (i.e. after the first) glyph contour had off curve start and end points, the phantom points were being added incorrectly. This has been resolved by tracking the start of the contour for the final points, as well, and using that (possibly calculated) finalised start point to complete the contour if required.