Add path handle data for designer.
Areas in an ISharpCanvas can now be registered with "handles" (points on a path that define the area geometry) which will be displayed in the designer. A new highlighting color has been introduced for these handles, and they have been incorporated into the Markup logic for paths, shapes, and groups.