Fix issue with origin tracking in card configuration parsing.
A possibly unnecessary refactor of the origin tracking to use a custom ParseOrigins class -- but the result is more readable. There was an issue with CardConfigEditorParser not accounting for origins from the original CardSetConfig parse, and therefore downstream tasks not being able to identify origins from the config, to find definitions. This has been solved in a slightly simplistic way (just appending the data), but this is probably the best approach, and the current code is overly complex and in need of refactoring (some of it might be redundant...?).