Unfinished: Evaluation types refactor.
Refactor is part complete, but it has become apparent that the new abstract EvaluationType class needs to be backed up by some kind of EvaluationContext (partially implemented). This is going to require some fairly major edits, and so this commit is a checkpoint before starting more major rewrites, in case something goes wrong.