Implement type literals in expressions.
Type literals can now be used to access static variables. This requires that the rules on variable names be made more strict, and that language keywords be defined. Card Definitions currently only place name reuse limitations on the first name (i.e. not on aliases), which may need updating later.