Add new DictionaryEvaluationType.
This type is also included in the MarkupEvaluationTypes parsing, and a Dictionary parser has been added to ValueParsing. However, it is currently backed by a Dictionary<,>, which does not guarantee that the order of insertion will match order of iteration, which is not desirable for this context. So, a SortedDictionary<,> (available in .NET 9) would be a better option (we're currently in .NET 6, so an update is required to get the generic SortedDictionary).