Introduce ShapeParams for shape construction.
This makes it easier to pass arbitrary but properly-typed build parameters into the MarkupPattern MakeShape method, and neatens up the code a touch. It is primarily intended to facilitate the creation of a Markup TitleStyle pattern type, which requires additional parameters (i.e. the box to wrap the title around).