Fix Evaluations.Parse bug for ComprehensionIfNode.
Because the provider variables were being collected before the comprehension source had its VariablePlaceholderNode instances replaced, this was causing an error when trying to determine the loop variable type. Now we simply list the provider variable names first, and only attempt the type lookup if the variable appears in the expression. This is perhaps less neat, but it does prevent this issue. The other option would be to try and provide information as to whether the provider is allowed for its own operands (on an operand-by-operand basis), but this seems even more awkward.

(Genuinely unsure how this has not occurred before. Presumably very little use was being made of ComprehensionIfNode up to now.)