Simplify IExpression<T> interface by creating non-generic parent IExpression.
The non-generic components of IExpression<T> have been moved into a non-generic parent interface IExpression, allowing for a nice refactor of some IExpression helper code (to use params). This change could also be beneficial when moving to source generator setup, as it will allow for easier testing of whether types are an instance of IExpression.