Add MakeContainer to ShapeFactory.
This simplifies the logic of MakeShape a little, by moving the
IContainerShape-specific construction logic into a separate helper
method. This method is also publicly accessible, and will be used by the
source generator setup to directly create containers when needed.

It is probably better, in the long run, to switch to having ITitleStyle
be a separate part of the drawing setup, and have widgets simply put
both "outline" and "title" objects in their builders, and then combine
the two themselves, but let's leave that for later.