Initial implementation of source generator for construction and documentation.
A source generator has been implemented which is intended to replace SharpFactory and the documentation pipeline. The various "factory" methods that are added onto partial types provide compile-time construction code for the predefined widget/shape/etc types, reducing the reliance on reflection (and so hopefully improving performance). The documentation is read directly from the source files and converted into compiled BuilderDetail/etc objects which are lazily instantiated at runtime when needed. Helper methods for implementing ITypeDetailsCollection/etc are also provided.

Additional Attributes have been created to support this, and minor changes made to the codebase where necessary to support this new setup.

Currently the pipeline compiles, and produces valid output. However, these are still many edge-cases unaccounted for, and most importantly the Incremental Source Generator pipeline makes overly-extensive use of Compilation objects, which will strongly negatively impact performance. Caching the necessary data early in the pipeline is preferable, and should be attempted. The generated code also needs to be properly wired-up to the existing build system, and then SharpFactory needs to be phased out (although Markup parsing still relies on it, which will have to be addressed, possibly necessitating some adjustments to EvaluationType internal logic).