Create TypeNameUtils.ReduceParameterTypeName.
This performs parameter type "reduction" on string names, without the need for a Compilation instance. This allows the Compilation instance to be removed from several places in the pipeline. Although this does not help with caching of pipeline steps here, as the Compilation is still required in later steps, it should hopefully speed up the earlier steps (as accessing the Compilation is an expensive operation?).

This is possibly a roundabout way of doing things, but it does make the code cleaner, removing a bunch of obtuse value passing between method calls.