Begin integration of source generated code into main codebase.
ShaprFactory, WidgetFactory, and CardSetConfigFactory have been refactored to use the source generated builder methods, and this appears to be working. This has necessitated a large number of additions to the FactoryGenerator code, and some adjustments to the main codebase. EvaluationType has had parsing behaviour added to reduce the dependency on reflection-based parsing.

SharpFactory and DocumentationGenerator are now mostly unused (with the exception of some methods in DocumentationGenerator which should probably be moved to a more appropriately named type now). However, Activator is still used in a number of places, which will need to be refactored.

This commit comes before any major removals of old behaviours (in particular, SharpFactory and DocumentationGenerator), which should now be possible.