Major refactor of type handling for display and Trimming.
DisplayType class added, and EvaluationType adjusted to remove DisplayType attribute. This required major rewrites of the editor serialisation code, and some other rewrites. Also removed code that did not play well with Assembly Trimming in SharpSheets, including some related to Type and tuple handling. SharpSheets should now be roughly Trimming compatible, although this has not been tested yet. Further testing is also required for the editor, as crashes and incorrect displays are still possible/likely. An additional source generator has been added to find and store IExpression<> types, to avoid the need for recursive interface finding to determine the expression type of an arbitrary type object. A bunch of other refactoring to support all this.