Allow for rotationg text fields in SharpSheets.
This was possible from GeboPDF, but had not been made explicitly available in SharpSheets. One caveat: The rotation is relative to the page, not calculated from the orientation of the current transform. Is this an issue in practice? Calculating the implied rotation is slightly non-trivial, and is not strictly necessary. Possible future improvement?