Add min-width/min-height attributes to Markup Div elements.
This allows for Markup patterns to have custom "minimum content size" values, which should improve the overall functionality of the Markup system. Currently these values override any other value. An alternative is to have the maximum value selected from the provided and calculated options, but this is not as in-keeping with the approach taken in the built-in SharpWidget implementation.