Reduce redendancy in ConstantNode GetRepresentation().
Make use of the EvaluationValue.ToEvaluationString() in ConstantNode. However, this presents a problem, where some other parts of the code are expecting string values to be represented by their raw content, but ToEvaluationString() should really put them in quote marks. The former case wins out at the moment, but it's an inconsistency, and should probably be dealt with at some point.