Rewrite GraphicsStream to trim some unnecessary operations.
The GraphicsStream now maintains a data structure of the graphics operations, rather than writing directly to a MemoryStream, so that it can make adjustments to the content when writing out, to remove unnecessary operations. In particular, unnecessary saving and restoring of state can be avoided this way.

Additional work is possible to remove unnecessary state changes which are not used before being overriden (this happens in several places in SharpSheets with font and size state updates, among others).