Refactor IShape hierarchy to allow for different IDetail draw method signature.
There is now an IDrawRectShape interface with a Draw method that just takes canvas and rect, with IDetail having it's own Draw with canvas, rect, and layout. This is much neater.

This also shows that splitting the hierarchy is perfectly doable for all of the dependent systems, which is the first step to changing the structure of other parts of the shape hierarchy as well.