Remove ITitleStyledBox and replace with ITitleStyle.
This has necessitated quite a lot of rewrites of certain areas, but overall leads to far fewer strange edge-cases. This has simplified code in several areas, and further simplifications are definitely possible. This state is mostly-working (in that there are no obvious bugs, but it has not been thoroughly tested).

Further cleanup and fine-tuning required.