Improvements to builder example creation.
The BuilderDetails class now has an Example (lazily evaluated) which can hold an instance of the built type. This allows for easier handling of creating example values for documentation, as now more complex example parameters can be constructed automatically and be made available when requested. To this end, an additional aspect of parsing Builder types for Default and Example documentation values has been implemented, where arguments to the builders can be provided in the generator attributes (e.g. "Simple(aspect: 2f)"). ShapeFactory has been adjusted to make use of these examples where available. WidgetFactory has not, as it currently injects additional example setup (gutter size, in particular), which is required to make things in the documentation look sensible at the moment.